requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.4.0
temporalio==1.17.0
qdrant-client==1.15.1
openai==1.102.0
anthropic==0.67.0
structlog==25.4.0
httpx==0.28.1

requests==2.32.4
